requirements.txt
Flask==2.3.3
Flask-WTF==1.2.1
WTForms==3.1.2
Flask-Limiter==3.5.0
cryptography==42.0.5
python-decouple==3.8
python-dateutil==2.9.0.post0
pillow==10.2.0
captcha==0.5.0
requests==2.31.0
pyotp==2.9.0
gunicorn==21.2.0
email-validator==2.2.0
